requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.10
dnspython==2.1.0
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
imageio-ffmpeg==0.4.5
mailjet-rest==1.3.4
multidict==5.2.0
numpy==1.22.0
pymongo==4.0.1
python-dotenv==0.19.2
requests==2.27.1
typing-extensions==4.0.1
urllib3==1.26.8
yarl==1.7.2
